add routing in search component
